// frontend/app/_generated/generated-abstract/cosmwasm-codegen/XionAdventuresHub.client.ts
export interface IXionAdventuresHubAppQueryClient {
  moduleId: string;
  accountPublicClient: AccountPublicClient;
  _moduleAddress: string | undefined;
  config: () => Promise<ConfigResponse>;
  nextTokenId: () => Promise<NextTokenIdResponse>;
  getAddress: () => Promise<string>;
}
export class XionAdventuresHubAppQueryClient implements IXionAdventuresHubAppQueryClient {
  accountPublicClient: AccountPublicClient;
  moduleId: string;
  _moduleAddress: string | undefined;

  constructor({
    accountPublicClient,
    moduleId
  }: {
    accountPublicClient: AccountPublicClient;
    moduleId: string;
  }) {
    this.accountPublicClient = accountPublicClient;
    this.moduleId = moduleId;
    this.config = this.config.bind(this);
    this.nextTokenId = this.nextTokenId.bind(this);
  }

  config = async (): Promise<ConfigResponse> => {
    return this._query(XionAdventuresHubQueryMsgBuilder.config());
  };
  nextTokenId = async (): Promise<NextTokenIdResponse> => {
    return this._query(XionAdventuresHubQueryMsgBuilder.nextTokenId());
  };
  getAddress = async (): Promise<string> => {
    if (!this._moduleAddress) {
      const address = await this.accountPublicClient.getModuleAddress({
        id: this.moduleId
      });

      if (address === null) {
        throw new Error(`Module ${this.moduleId} not installed`);
      }

      this._moduleAddress = address;
    }

    return this._moduleAddress!;
  };
  _query = async (queryMsg: QueryMsg): Promise<any> => {
    return this.accountPublicClient.queryModule({
      moduleId: this.moduleId,
      moduleType: "app",
      queryMsg
    });
  };
}
export interface IXionAdventuresHubAppClient extends IXionAdventuresHubAppQueryClient {
  accountWalletClient: AccountWalletClient;
  ibcTransfer: (params: CamelCasedProperties<Extract<ExecuteMsg, {
    ibc_transfer: unknown;
  }>["ibc_transfer"]>, fee_?: number | StdFee | "auto", memo_?: string, funds_?: Coin[]) => Promise<ExecuteResult>;
  mint: (params: CamelCasedProperties<Extract<ExecuteMsg, {
    mint: unknown;
  }>["mint"]>, fee_?: number | StdFee | "auto", memo_?: string, funds_?: Coin[]) => Promise<ExecuteResult>;
  modifyMetadata: (params: CamelCasedProperties<Extract<ExecuteMsg, {
    modify_metadata: unknown;
  }>["modify_metadata"]>, fee_?: number | StdFee | "auto", memo_?: string, funds_?: Coin[]) => Promise<ExecuteResult>;
}
export class XionAdventuresHubAppClient extends XionAdventuresHubAppQueryClient implements IXionAdventuresHubAppClient {
  accountWalletClient: AccountWalletClient;

  constructor({
    accountPublicClient,
    accountWalletClient,
    moduleId
  }: {
    accountPublicClient: AccountPublicClient;
    accountWalletClient: AccountWalletClient;
    moduleId: string;
  }) {
    super({
      accountPublicClient,
      moduleId
    });
    this.accountWalletClient = accountWalletClient;
    this.ibcTransfer = this.ibcTransfer.bind(this);
    this.mint = this.mint.bind(this);
    this.modifyMetadata = this.modifyMetadata.bind(this);
  }

  ibcTransfer = async (params: CamelCasedProperties<Extract<ExecuteMsg, {
    ibc_transfer: unknown;
  }>["ibc_transfer"]>, fee_: number | StdFee | "auto" = "auto", memo_?: string, funds_?: Coin[]): Promise<ExecuteResult> => {
    return this._execute(XionAdventuresHubExecuteMsgBuilder.ibcTransfer(params), fee_, memo_, funds_);
  };
  mint = async (params: CamelCasedProperties<Extract<ExecuteMsg, {
    mint: unknown;
  }>["mint"]>, fee_: number | StdFee | "auto" = "auto", memo_?: string, funds_?: Coin[]): Promise<ExecuteResult> => {
    return this._execute(XionAdventuresHubExecuteMsgBuilder.mint(params), fee_, memo_, funds_);
  };
  modifyMetadata = async (params: CamelCasedProperties<Extract<ExecuteMsg, {
    modify_metadata: unknown;
  }>["modify_metadata"]>, fee_: number | StdFee | "auto" = "auto", memo_?: string, funds_?: Coin[]): Promise<ExecuteResult> => {
    return this._execute(XionAdventuresHubExecuteMsgBuilder.modifyMetadata(params), fee_, memo_, funds_);
  };
  _execute = async (msg: ExecuteMsg, fee_: number | StdFee | "auto" = "auto", memo_?: string, funds_?: Coin[]): Promise<ExecuteResult> => {
    const signingCwClient = await this.accountWalletClient.getSigningCosmWasmClient();
    const sender = await this.accountWalletClient.getSenderAddress();
    const moduleMsg: AppExecuteMsg<ExecuteMsg> = AppExecuteMsgFactory.executeApp(msg);
    return await signingCwClient.execute(sender, await this.getAddress(), moduleMsg, fee_, memo_, funds_);
  };
}